requirements.txt
opencv-python==4.5.5.62
paho-mqtt==1.6.1